Added #all query
Created an initial query with the all query that takes no parameters and
returns all the documents in a collection.

Refactored some common setup code in the cucumis step definitions into
it's own file. Also refactored the cucumis TestItem model into it's own
file and created a models directory to hold them.

Extended the should matchers to have one that looks for objects in an
array that match a single given key/value pair.

Removed the underscore iterator on keys in an object and used the
appropriate coffee script construct instead.

Improved the #save spec by calling through to the real tableize method.